+++ Gestor-Financiero/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "java.debug.settings.onBuildFailureProceed": true
+}
